# public_html/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your Awesome Modules's API",
        "contact": {
            "name": "Developer Team",
            "email": "developers@module.com"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/salary": {
            "get": {
                "tags": [
                    "MongoDB Salary"
                ],
                "operationId": "/salary/",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "MongoDB Salary"
                ],
                "summary": "Create Salary",
                "description": "Create Salary",
                "operationId": "create salary",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "5645565s4d0sd4546654",
                                        "type": "string"
                                    },
                                    "salary": {
                                        "description": "7000000",
                                        "type": "integer"
                                    },
                                    "date": {
                                        "description": "2021-03-30",
                                        "type": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/salary/{id}": {
            "get": {
                "tags": [
                    "MongoDB Salary"
                ],
                "summary": "Get salary information",
                "description": "Returns salary data",
                "operationId": "getSalaryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Salary id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "MongoDB Salary"
                ],
                "summary": "Update Salary",
                "description": "Update Salary",
                "operationId": "salary id",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "5645565s4d0sd4546654",
                                        "type": "string"
                                    },
                                    "salary": {
                                        "description": "7000000",
                                        "type": "integer"
                                    },
                                    "date": {
                                        "description": "2021-03-30",
                                        "type": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/salary-user/{id}": {
            "get": {
                "tags": [
                    "MongoDB Salary"
                ],
                "summary": "Get salary by user id information",
                "description": "Returns salary data",
                "operationId": "getSalaryByUserId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/Salary/{id}": {
            "delete": {
                "tags": [
                    "MongoDB Salary"
                ],
                "summary": "Delete existing salary data",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteSalary",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Salary id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}